task-97: create separate mock api for storybook
