+++ src/js/api_clients/index.ts
@@ -1,2 +1,4 @@
 export * from "./createNewEvent";
 export * from "./listPoint";
+export * from "./api_members";
+export * from "./storybook";
