+++ src/js/api_clients/storybook/index.ts
@@ -1,0 +1,3 @@
+export * from "./listPoints/commonListPoints";
+export * from "./listPoints/privateListPoints";
+export * from "./listPoints/takenListPoints";
